Guess my favorite number
The code will choose a random number between 1 and 100 and you have to guess it using Binary search. Every time you guess a number, the system will tell you whether the actual number is greater or less than your guessed number. Guess the number in the minimum number of steps.